Added gradients and text and font changes.

+++ App/NLSSQLAPI.m
@@ -563,7 +563,7 @@
         return;
     }];
     
-    NSLog(@"count result: %ld", count);
+    NSLog(@"count result: %ld", (unsigned long)count);
     return count;
     
 }
